% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control.R
\name{em_control}
\alias{em_control}
\title{Configure the EM algorithm}
\usage{
em_control(
  max_iter = 1000L,
  tolerance = 1e-05,
  min_weight = 1e-05,
  verbose = FALSE
)
}
\arguments{
\item{max_iter}{Maximum number of iterations.}

\item{tolerance}{Convergence tolerance (convergence assumed when delta in loglikelihood falls below tolerance).}

\item{min_weight}{Minimum permitted component weight.}

\item{verbose}{Whether to print iteration progress.}
}
\value{
A list of EM control settings.
}
\description{
Configure the EM algorithm
}
