% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit-exp-normal.R
\name{fit_exp_normal}
\alias{fit_exp_normal}
\title{Fit one dimensional exponential-normal density using EM algorithm}
\usage{
fit_exp_normal(data, start = NULL, control = list())
}
\arguments{
\item{data}{One dimensional dataset to fit mixture into.}

\item{start}{Optional list of initial approximations to parameters.}

\item{control}{List that controls how EM is fitted (tolerance, max iterations)}
}
\value{
A list of fitted parameters, weights, component bayes_probilities, iterations, convergence, loglikelihood and deltas
}
\description{
Fit one dimensional exponential-normal density using EM algorithm
}
