% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_exp_normal.R
\name{gen_exp_norm}
\alias{gen_exp_norm}
\title{Generate one dimensional exponential-normal density}
\usage{
gen_exp_norm(size, parameters)
}
\arguments{
\item{size}{Size of the dataset to generate.}

\item{parameters}{List of parameters.}
}
\value{
Vector generated from the mixture.
}
\description{
Generate one dimensional exponential-normal density
}
