% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_exp_normal.R
\name{plot_exp_norm}
\alias{plot_exp_norm}
\title{Plot output of EM fit to 1-dim exp-normal density using EM.}
\usage{
plot_exp_norm(em_exp_norm_list)
}
\arguments{
\item{em_exp_norm_list}{Output of fit_exp_normal()}
}
\description{
Plot output of EM fit to 1-dim exp-normal density using EM.
}
