% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_exp_normal_parameters}
\alias{validate_exp_normal_parameters}
\title{Validate that parameters for exp-normal density are specified correctly}
\usage{
validate_exp_normal_parameters(parameters)
}
\arguments{
\item{parameters}{List with parameters}
}
\value{
parameters as vector
}
\description{
Validate that parameters for exp-normal density are specified correctly
}
