% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-exp-normal.R
\name{plot.em_exp_normal_fit}
\alias{plot.em_exp_normal_fit}
\title{Plot output of EM fit to 1-dim exp-normal density using EM.}
\usage{
\method{plot}{em_exp_normal_fit}(x, ...)
}
\arguments{
\item{x}{A fitted exponential-normal mixture model.}

\item{...}{Additional graphical arguments.}
}
\value{
The fitted model, invisibly.
}
\description{
Plot output of EM fit to 1-dim exp-normal density using EM.
}
